requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
diskcache==5.6.3
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
llama_cpp_python==0.3.14
MarkupSafe==3.0.2
narwhals==2.0.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
plotly==6.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
